feat: wagmi provider
