+++ frontend/index.tsx
@@ -1,39 +1,16 @@
 import React from "react";
-import ReactDOM from "react-dom/client";
+import { createRoot } from "react-dom/client";
 import "@/i18n";
 
 import "./index.css";
 import "@rainbow-me/rainbowkit/styles.css";
 
 import App from "./App";
-import { RainbowKitProvider, darkTheme } from "@rainbow-me/rainbowkit";
-
-import { chains, wagmiConfig } from "@/config/wagmi.config";
-import { WagmiConfig } from "wagmi";
-
-import { canisterId, idlFactory } from "@/candid/ic_siwe_provider";
-
 import reportWebVitals from "./reportWebVitals";
-import { SiweIdentityProvider } from "@/siwe";
 
-ReactDOM.createRoot(document.getElementById("root")!).render(
+createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <WagmiConfig config={wagmiConfig}>
-      <RainbowKitProvider
-        chains={chains}
-        modalSize="compact"
-        theme={darkTheme({
-          accentColor: "#7b3fe4",
-          accentColorForeground: "white",
-          borderRadius: "large",
-          overlayBlur: "none",
-        })}
-      >
-        <SiweIdentityProvider canisterId={canisterId} idlFactory={idlFactory}>
-          <App />
-        </SiweIdentityProvider>
-      </RainbowKitProvider>
-    </WagmiConfig>
+    <App />
   </React.StrictMode>,
 );
